东方财富-简单的股票爬虫程序1.2

+++ src/main/java/com/stock/spider/controller/SpiderController.java
@@ -67,8 +67,8 @@ public class SpiderController {
 
     @RequestMapping("/stock/{code}")
     @ResponseBody
-    public Map<Double, String> stock(@PathVariable("code") String code) {
-        Map<Double, String> stock = stockService.stock(code);
+    public Map<Integer, String> stock(@PathVariable("code") String code) {
+        Map<Integer, String> stock = stockService.stock(code);
         return stock;
     }
 }
