+++ src/main/java/com/stock/spider/service/StockService.java
@@ -7,5 +7,5 @@ import java.util.Map;
  * 这里只找出股票，具体的股票信息靠手动去分析，什么基本面啊技术面啊筹码分布啊
  */
 public interface StockService {
-    Map<Double, String> stock(String industryCode);
+    Map<Integer, String> stock(String industryCode);
 }
